+++ pluto/test/test_rx_debug.py
@@ -47,7 +47,7 @@ def main() -> None:
             filtered = np.convolve(samples, h_rrc, mode="same")
 
             # Get correlation metrics from synchronizer
-            result = sync.detect_preamble(filtered, SAMPLE_RATE)
+            result = sync.detect_preamble(filtered, np.float32(SAMPLE_RATE))
 
             # Also compute raw correlation manually to see values
             short_zc = sync.zc_short
@@ -55,13 +55,13 @@ def main() -> None:
 
             if len(filtered) >= short_len:
                 # Correlate with short ZC
-                corr = np.abs(np.correlate(filtered[: short_len * 4], np.repeat(short_zc, SPS), mode="valid"))
+                corr = np.abs(np.correlate(filtered[: short_len * 4], np.repeat(short_zc, SPS).astype(np.complex64), mode="valid"))
                 max_corr = np.max(corr) if len(corr) > 0 else 0
 
                 # Normalize
                 sig_power = np.mean(np.abs(filtered[: short_len * 4]) ** 2)
                 ref_power = np.mean(np.abs(np.repeat(short_zc, SPS)) ** 2)
-                max_corr / (np.sqrt(sig_power * ref_power) * len(short_zc) * SPS + 1e-12)
+                max_corr / (np.sqrt(sig_power * ref_power) * len(short_zc) * SPS + np.float32(1e-12))
 
 
                 if result.success:
